Move the code from main_test.go to main.go
